5.2 RPPM Trinket stuff
- HT Renataki's has the right value now.
- Added the "spell haste buff" to RPPM trinkets. There's still a lack
of empirical data, but I'm just going to go with the general attitude
among the rest of the theorycrafters.

TODO:
- Still need to verify the right value for Bad Juju and Talisman. But
these are my best guesses, I suppose there's no point in holding onto
them still if they're close enough.

Unless someone else determines a major bug, and it isn't just
speculation on their part, this will be my last update for 5.2. Any
commits after this should be experimental and/or for 5.3.