5.4 Sub Tweaks (Unoptimized and Experimental)
- Fixed energy consumption when opening from stealth.
- Set ShD pooling to use an adv_param, default to 10 below max.
- Tweaking swing resets.
- Slightly lower Shadowmeld usage.
- Ambush from CDs better accounts for the opening sequence of events.
That is, it accounts for the fact that not every CD is blown at the
pull.
- Shadow Dance now relies on a loop to gauge APS. Better handles
smaller time intervals.
- FW uptime better handles opening sequence of events like Ambush now
does.
- Autoattack's FW weight was fixed to support AoC.
- t16 4pc Accounts for overlap.

- THINGS TO NOTE: Ambush damage might still be a little high? Either
the ambush buff didn't go through, the FW modifier is broken still,
people aren't pooling before ShD properly (and I'm struggling to see
the failures), or there's still a bug with the APS. May be more than
one issue.