6.0 Fixes
- Hemo dot crit hotfix.
- Added attack speed scaling support for damaging procs.
- Set the Shattered Hand enchant to scale with attack speed.

NOTE: Shatt Hand is most likely bugged. All signs point to this enchant
scaling with attack speed in addition to haste. This is a rogue centric
issue due to Slice and Dice being one of few abilities like it
remaining.

The data is not conclusive, but all data currently indicates this
pattern.